+++ foodie-be/app/Http/Controllers/RestaurantController.php
@@ -2,12 +2,15 @@
 
 namespace App\Http\Controllers;
 
+use Symfony\Component\HttpKernel\Exception\BadRequestHttpException;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
 use App\Exceptions\InternalErrorException;
 use App\Http\Requests\CreateMealRequest;
 use App\Http\Requests\CreateOrderRequest;
+use App\Http\Requests\CreateRestaurantRequest;
 use App\Http\Requests\UpdateRestaurantRequest;
 use App\Interfaces\IRestaurantService;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
+use App\Models\Restaurant;
 use Exception;
 use Log;
 
@@ -24,6 +27,7 @@ class RestaurantController extends Controller
     {
         $this->restaurant_service = $restaurant_service;
     }
+    
     /**
      * Get all restaurants.
      *
@@ -33,7 +37,7 @@ class RestaurantController extends Controller
     public function index()
     {
         try {
-            return $this->restaurant_service->getAllRestaurants();
+            return $this->restaurant_service->getActiveRestaurants();
         } catch (Exception $e) {
             Log::error('Get restaurants, Exception', ['error' => $e->getMessage()]);
             throw new InternalErrorException();
@@ -61,6 +65,33 @@ class RestaurantController extends Controller
             }
 
             Log::error('Get restaurant by id, Exception', ['error' => $e->getMessage()]);
+            throw new InternalErrorException();
+        }
+    }
+    
+    /**
+     * Create a new restaurant.
+     * 
+     * @param App\Http\Requests\CreateRestaurantRequest $request
+     * 
+     * @throws InternalErrorException 
+     * @return App\Models\Restaurant $restaurant
+     */
+    public function create(CreateRestaurantRequest $request)
+    {
+        $this->authorize('create', Restaurant::class);
+
+        try {
+            $payload = $request->only([
+                'name',
+                'food_type',
+                'description',
+            ]);
+            $owner = $request->user();
+
+            return $this->restaurant_service->createRestaurant($payload, $owner);
+        } catch (Exception $e) {
+            Log::error('Create a new restaurant, Exception', ['error' => $e->getMessage()]);
             throw new InternalErrorException();
         }
     }
@@ -101,6 +132,8 @@ class RestaurantController extends Controller
      */
     public function createMeal(CreateMealRequest $request, $id)
     {
+        $this->authorize('createMeal', $this->show($id));
+
         try {
             $payload = $request->only([
                 'name',
@@ -130,8 +163,10 @@ class RestaurantController extends Controller
      * @throws ModelNotFoundException 
      * @return App\Models\Restaurant $restaurant
      */
-    public function updateRestaurant(UpdateRestaurantRequest $request, $id)
+    public function update(UpdateRestaurantRequest $request, $id)
     {
+        $this->authorize('update', $this->show($id));
+
         try {
             $payload = $request->only([
                 'name',
@@ -160,10 +195,12 @@ class RestaurantController extends Controller
      * @throws ModelNotFoundException 
      * @return void
      */
-    public function deleteRestaurant($id)
+    public function destroy($id)
     {
+        $this->authorize('delete', $this->show($id));
+
         try {
-            $this->restaurant_service->delete($id);
+            return $this->restaurant_service->delete($id);
         } catch (Exception $e) {
             if($e instanceof ModelNotFoundException) {
                 Log::warning('Delete restaurant by id, ModelNotFoundException', ['error' => $e->getMessage()]);
@@ -186,6 +223,8 @@ class RestaurantController extends Controller
      */
     public function getRestaurantOrders($id)
     {
+        $this->authorize('getRestaurantOrders', $this->show($id));
+
         try {
             return $this->restaurant_service->getRestaurantOrders($id);
         } catch (Exception $e) {
@@ -207,15 +246,23 @@ class RestaurantController extends Controller
      * 
      * @throws InternalErrorException 
      * @throws ModelNotFoundException 
+     * @throws BadRequestException 
      * @return Collection $orders
      */
     public function createOrder(CreateOrderRequest $request, $id)
     {
+        $this->authorize('createOrder', $this->show($id));
+
         try {
             return $this->restaurant_service->createOrder($id, $request->user(), $request['mealIds']);
         } catch (Exception $e) {
             if($e instanceof ModelNotFoundException) {
                 Log::warning('Create a new order for the restaurant, ModelNotFoundException', ['error' => $e->getMessage()]);
+                throw $e;
+            }
+
+            if($e instanceof BadRequestHttpException) {
+                Log::warning('Create a new order for the restaurant, BadRequestHttpException', ['error' => $e->getMessage()]);
                 throw $e;
             }
