+++ foodie-be/app/Http/Controllers/UserController.php
@@ -4,10 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Exceptions\InternalErrorException;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\CreateRestaurantRequest;
 use App\Http\Requests\UpdateUserRequest;
 use App\Interfaces\IUserService;
-use Illuminate\Http\Request;
 use Exception;
 use Auth;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
@@ -45,6 +43,8 @@ class UserController extends Controller
      */
     public function index()
     {
+        $this->authorize('viewAll', User::class);
+
         try {
             return $this->user_service->getAllUsers();
         } catch (Exception $e) {
@@ -65,7 +65,7 @@ class UserController extends Controller
     public function show($id)
     {
         try {
-            return $this->user_service->getUser($id);
+            $user =  $this->user_service->getUser($id);
         } catch (Exception $e) {
             if($e instanceof ModelNotFoundException) {
                 Log::warning('Get user by id, ModelNotFoundException', ['error' => $e->getMessage()]);
@@ -75,6 +75,10 @@ class UserController extends Controller
             Log::error('Get user by id, Exception', ['error' => $e->getMessage()]);
             throw new InternalErrorException();
         }
+
+        $this->authorize('view', $user);
+
+        return $user;
     }
 
     /**
@@ -87,8 +91,10 @@ class UserController extends Controller
      * @throws ModelNotFoundException 
      * @return App\Models\User $user
      */
-    public function updateUser(UpdateUserRequest $request, $id)
+    public function update(UpdateUserRequest $request, $id)
     {
+        $this->authorize('update', $this->show($id));
+
         try {
             $payload = $request->only([
                 'name',
@@ -117,8 +123,10 @@ class UserController extends Controller
      * @throws ModelNotFoundException 
      * @return void
      */
-    public function deleteUser($id)
+    public function destroy($id)
     {
+        $this->authorize('delete', $this->show($id));
+
         try {
             $this->user_service->delete($id);
         } catch (Exception $e) {
@@ -128,48 +136,6 @@ class UserController extends Controller
             }
 
             Log::error('Delete user by id, Exception', ['error' => $e->getMessage()]);
-            throw new InternalErrorException();
-        }
-    }
-    
-    /**
-     * Create a new restaurant.
-     * 
-     * @param App\Http\Requests\CreateRestaurantRequest $request
-     * 
-     * @throws InternalErrorException 
-     * @return App\Models\Restaurant $restaurant
-     */
-    public function createRestaurant(CreateRestaurantRequest $request)
-    {
-        try {
-            $payload = $request->only([
-                'name',
-                'food_type',
-                'description',
-            ]);
-            $owner = $request->user();
-
-            return $this->user_service->createRestaurant($payload, $owner);
-        } catch (Exception $e) {
-            Log::error('Create a new restaurant, Exception', ['error' => $e->getMessage()]);
-            throw new InternalErrorException();
-        }
-    }
-
-    /**
-     * Get the user's orders
-     * 
-     * @param Illuminate\Http\Request $request
-     * @throws InternalErrorException 
-     * @return Collection $orders
-     */
-    public function getUserOrders(Request $request)
-    {
-        try {
-            return $this->user_service->getUserOrders($request->user());
-        } catch (Exception $e) {
-            Log::error(' Get all orders for the user, Exception', ['error' => $e->getMessage()]);
             throw new InternalErrorException();
         }
     }
